Feature for sitemapindex
The ReadMe has been modified to explain the new feature.

The start option has been added. It starts automatically the element
urlset. When set to false, it disable the startElement and needs a
manual declaration with the method startElement.
This method has been added to the SitemapGenerator.

The methods addEntry and addSitemap now returns the Class to have the
posibility to chain the entries together. (See the example into the
readme.)

A test has been added to the feature. I am not familiar yet with the
library used.
As PHPunit and PHPspecs are not implemented, I did not add those tests.